✅ refactor fixture usage
