+++ tests/integration/fixtures/page.ts
@@ -1,10 +1,10 @@
 import { test as base } from '@playwright/test';
 
-export const baseTest = base.extend({
+export const test = base.extend({
   page: async ({ page }, use) => {
     // ---------- BEFORE EACH TEST ----------
     await page.goto('http://localhost:5173/');
-    await page.waitForSelector('[data-app-status="ready"]', { state: 'attached' });
+    await page.locator('[data-app-status="ready"]').waitFor({ state: 'attached' });
     await use(page);
   },
 });
